requirements.txt
django
django-organizations
django-environ
psycopg2-binary
django-extensions
djangorestframework
markdown
django-filter
